Edit profile and bug fixes

+++ web/pages/api/workout/add.ts
@@ -27,7 +27,7 @@ export default async function handler(
                 return res.status(400).json({message: "Invalid calories burned"})
             }
 
-            if(!duration || isNaN(parseInt(duration)) || parseInt(duration) <= 0) {
+            if(!duration || isNaN(parseFloat(duration)) || parseFloat(duration) <= 0) {
                 return res.status(400).json({message: "Invalid duration"})
             }
 
@@ -65,7 +65,7 @@ export default async function handler(
                         }
                     },
                     caloriesBurned: parseInt(caloriesBurned),
-                    duration: parseInt(duration)
+                    duration: parseFloat(duration)
                 },
                 include: {
                     workoutCategory: true,
